+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+- v1.02 (2022-08-05)
+  - Now works with offline graphs
+  - Added extension settings
+  - Extension setting to set max links tracked, 5-100
+  - Extension setting to set minimum and maximum width of the breadcrumbs container (minimum width is for when the right sidebar is open)
 - v1.01 (2022-08-03)
   - Added functionality for tracking visits to user's "Graph Overview" and "All Pages"
 - v1.0 (2022-08-02)